+++ monticore-generator/src/main/groovy/de/monticore/MCPlugin.groovy
@@ -13,7 +13,6 @@ public class MCPlugin implements Plugin<Project> {
   
   public void apply(Project project) {
     project.ext.MCTask = de.monticore.MCTask
-    project.ext.GroovyTask = de.monticore.GroovyTask
     project.configurations.create("grammar")
   }
 }