+++ monticore-generator/src/main/java/de/monticore/dstlgen/util/DetermineOptionalRuleComponentsVisitor.java
@@ -67,11 +67,6 @@ public class DetermineOptionalRuleComponentsVisitor implements
   }
 
   @Override
-  public void traverse(ASTEof node) {
-
-  }
-
-  @Override
   public void traverse(ASTSemanticpredicateOrAction node) {
 
   }