+++ monticore-grammar/src/test/java/de/monticore/grammar/symboltable/MontiCoreGrammarSymbolTableCreatorTest.java
@@ -290,9 +290,9 @@ public class MontiCoreGrammarSymbolTableCreatorTest {
 
     Map<String, Collection<String>> tokenModes = grammar.getTokenModesWithInherited();
     assertEquals(3, tokenModes.size());
-    assertEquals(3, tokenModes.get(MCGrammarSymbol.DEFAULT_MODE).size());
-    assertEquals(1, tokenModes.get("Bla").size());
-    assertEquals(2, tokenModes.get("Foo").size());
+    assertEquals(4, tokenModes.get(MCGrammarSymbol.DEFAULT_MODE).size());
+    assertEquals(1, tokenModes.get("FOO_MODE").size());
+    assertEquals(1, tokenModes.get("BLA_MODE").size());
   }
 
   private int countExternalProd(MCGrammarSymbol grammar) {