+++ monticore-test/monticore-grammar-it/src/main/java/mc/testcd4analysis/_symboltable/TestCD4AnalysisSTCompleteTypes.java
@@ -65,10 +65,8 @@ public class TestCD4AnalysisSTCompleteTypes implements TestCD4AnalysisVisitor2 {
     CDMethOrConstrSymbol methodSymbol = astMethod.getSymbol();
     methodSymbol.setIsStatic(astMethod.getModifier().isStatic());
     setReturnTypeOfMethod(methodSymbol, astMethod);
-    if (!astMethod.getCDParameterList().isEmpty()) {
-      if (astMethod.getCDParameter(astMethod.getCDParameterList().size() - 1).isEllipsis()) {
-        methodSymbol.setIsEllipsis(true);
-      }
+    if (!astMethod.getCDParameterList().isEmpty() && astMethod.getCDParameter(astMethod.getCDParameterList().size() - 1).isEllipsis()) {
+      methodSymbol.setIsEllipsis(true);
     }
   }